requirements.txt
pandas>=2.0
numpy>=1.25
requests>=2.31
statsmodels>=0.14
matplotlib>=3.8
PyYAML>=6.0
